requirements.txt
# Email Inbox Organizer - Requirements
# Python 3.12 Compatible Version

# Core Dependencies
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# LLM & AI Framework
langchain>=0.1.14
langchain-core>=0.1.33
langchain-openai>=0.1.1
langchain-google-genai>=0.0.10
langgraph>=0.0.28

# Data Processing
pandas>=2.1.4
numpy>=1.26.0

# Frontend
streamlit>=1.32.2
streamlit-option-menu>=0.3.0

# Utilities
pydantic>=2.5.3
requests>=2.31.0

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.22.1
black>=23.12.1
flake8>=6.1.0
mypy>=1.8.0

# Optional: For enhanced features
# google-cloud-speech==2.24.0  # For voice integration
# librosa==0.10.0  # For audio processing
